refactor(Product at el): after separate packages, move entity, service, etc about Product
