+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,6 +1,6 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.product.domain.MenuProduct;
 
 import java.math.BigDecimal;
 import java.util.List;
